requirements.txt
asgiref==3.9.1
dj-database-url==3.0.1
dj-email-url==1.0.6
Django==5.2.4
django-allauth==65.10.0
django-cache-url==3.4.5
django-debug-toolbar==5.2.0
django-extensions==4.1
environs==14.2.0
marshmallow==4.0.0
psycopg==3.2.9
psycopg-binary==3.2.9
python-dotenv==1.1.1
sqlparse==0.5.3
typing_extensions==4.14.1
